+++ app/lib/instructions.ts
@@ -1,0 +1,22 @@
+export const instructions = [
+  {
+    instructionTitle: "Select the accidental:",
+    instructionText:
+      "Choose the flat (b) or sharp (#) button to select the type of accidental you need to add.",
+  },
+  {
+    instructionTitle: "Place it on the staff:",
+    instructionText:
+      "After selecting the accidental, click on the exact line or space on the staff where the accidental belongs.",
+  },
+  {
+    instructionTitle: "Erase a note or accidental:",
+    instructionText:
+      "If you need to erase a note or an accidental, first select the “Eraser” button, then click on the note or accidental you wish to erase.",
+  },
+  {
+    instructionTitle: "Clear a measure:",
+    instructionText:
+      "If you wish to clear an entire measure, select the “Clear Measure” button and the contents of the measure will be erased.",
+  },
+];
